Refactor addon into package
